+++ examples/ISR_Modify_PWM/ISR_Modify_PWM.ino
@@ -70,19 +70,19 @@ bool IRAM_ATTR TimerHandler(void * timerNo)
 uint32_t PWM_Pin    = LED_BUILTIN;
 
 // You can assign any interval for any timer here, in Hz
-double PWM_Freq1   = 1.0f;
+float PWM_Freq1   = 1.0f;
 // You can assign any interval for any timer here, in Hz
-double PWM_Freq2   = 2.0f;
+float PWM_Freq2   = 2.0f;
 
 // You can assign any interval for any timer here, in microseconds
-double PWM_Period1 = 1000000.0 / PWM_Freq1;
+uint32_t PWM_Period1 = 1000000 / PWM_Freq1;
 // You can assign any interval for any timer here, in microseconds
-double PWM_Period2 = 1000000.0 / PWM_Freq2;
+uint32_t PWM_Period2 = 1000000 / PWM_Freq2;
 
 // You can assign any duty_cycle for any PWM here, from 0-100
-double PWM_DutyCycle1  = 10.0;
+float PWM_DutyCycle1  = 10.0;
 // You can assign any duty_cycle for any PWM here, from 0-100
-double PWM_DutyCycle2  = 90.0;
+float PWM_DutyCycle2  = 90.0;
 
 // Channel number used to identify associated channel
 int channelNum;
@@ -111,7 +111,7 @@ void setup()
 
   // Just to demonstrate, don't use too many ISR Timers if not absolutely necessary
   // You can use up to 16 timer for each ISR_PWM
-  //void setPWM(uint32_t pin, uint32_t frequency, uint32_t dutycycle
+  //void setPWM(uint32_t pin, float frequency, float dutycycle
   // , timer_callback_p StartCallback = nullptr, timer_callback_p StopCallback = nullptr)
   Serial.print(F("Using PWM Freq = ")); Serial.print(PWM_Freq1); Serial.print(F(", PWM DutyCycle = ")); Serial.println(PWM_DutyCycle1);
 
@@ -137,8 +137,8 @@ void changePWM()
 {
   static uint8_t count = 1;
 
-  double PWM_Freq;
-  uint32_t PWM_DutyCycle;
+  float PWM_Freq;
+  float PWM_DutyCycle;
 
   if (count++ % 2)
   {
